+++ lib/src/enum/app_enums.dart
@@ -12,12 +12,3 @@ enum SelectionType {
 /// * [WrapType.scroll]: scroll the selected items horizontally
 /// * [WrapType.wrap]: wrap the selected items in both directions
 enum WrapType { scroll, wrap }
-
-/// [RequestMethod]
-/// RequestMethod enum for the request method of the dropdown items.
-/// * [RequestMethod.get]: get request
-/// * [RequestMethod.post]: post request
-/// * [RequestMethod.put]: put request
-/// * [RequestMethod.delete]: delete request
-/// * [RequestMethod.patch]: patch request
-enum RequestMethod { get, post, put, patch, delete }
